expose only std::span<..> and [Consuamble. Producable]Span to processBulk(...)
* hides the internal std::vector/std:array implementation to indicate more clearly to the user that the user-code does only get a non-owning reference to the input and output port data.

* minor compiler warning fixes (now there should be only the `Graph`-constructor one remain to be fixed)

+++ algorithm/include/gnuradio-4.0/algorithm/ImChart.hpp
@@ -269,8 +269,8 @@ public:
         _n_datasets++;
         if (axis_min_x == axis_max_x) {
             const auto [minX, maxX] = std::ranges::minmax_element(xValues);
-            axis_min_x              = *minX;
-            axis_max_x              = *maxX;
+            axis_min_x              = static_cast<double>(*minX);
+            axis_max_x              = static_cast<double>(*maxX);
             if (axis_min_x == axis_max_x) {
                 axis_max_x += 1.0; // safe fall back for x-range
             }
